ml/requirements.txt
# ── Core ML ──
xgboost==2.1.3
scikit-learn==1.6.1
pandas==2.2.3
numpy==1.26.4
optuna==4.1.0
joblib==1.4.2

# ── Data formats ──
pyarrow==18.1.0
pyyaml==6.0.2
python-dotenv==1.0.1

# ── API & Data Acquisition ──
requests==2.32.3
beautifulsoup4==4.12.3

# ── Geospatial ──
geopy==2.4.1
shapely==2.0.6

# ── gRPC ──
grpcio==1.68.1
grpcio-tools==1.68.1
grpcio-health-checking==1.68.1
protobuf==5.29.2

# ── Server ──
fastapi==0.115.6
uvicorn[standard]==0.34.0

# ── Evaluation & Plotting ──
matplotlib==3.10.0
seaborn==0.13.2

# ── Logging ──
structlog==24.4.0
